sort stack  of 3 elements

+++ lib/function_from_lib/ft_atoi.c
@@ -6,17 +6,18 @@
 /*   By: iaskour <iaskour@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/10/26 10:38:48 by iaskour           #+#    #+#             */
-/*   Updated: 2024/11/12 17:14:03 by iaskour          ###   ########.fr       */
+/*   Updated: 2025/02/03 15:26:35 by iaskour          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "../../Mandatory/push_swap.h"
+#include <stdio.h>
 
-int	ft_atoi(const char *str)
+long	ft_atoi(const char *str)
 {
 	long	num;
 	long	holder;
-	int		muns;
+	long	muns;
 
 	holder = 0;
 	while (*str != '\0' && ((*str >= 9 && *str <= 13) || (*str == 32)))
